requirements.txt
argparse
requests
python-whois
dnspython
shodan
